+++ src/main/kotlin/icu/windea/pls/config/cwt/config/CwtConfigGroupInfo.kt
@@ -25,27 +25,27 @@ data class CwtConfigGroupInfo(
 	val iconPathExpressions = mutableSetOf<String>()
 	
 	/**
-	 * @see CwtDataType.TypeExpressionString
+	 * @see CwtDataType.TemplateExpression
 	 */
-	val typeExpressionStringLinks = mutableListOf<LocationLink>()
-	//val typeExpressionStringLinks = mutableMapOf<String, MutableList<LocationLink>>() //typeExpression - links
+	val templateExpressionLinks = mutableListOf<LocationLink>()
+	//val templateExpressionLinks = mutableMapOf<String, MutableList<LocationLink>>() //typeExpression - links
 	
 	fun acceptConfigExpression(configExpression: CwtDataExpression) {
 		when(configExpression.type) {
-			CwtDataType.TypeExpressionString -> {
-				val link = configExpression.extraValue?.castOrNull<Pair<String, String>>()?.let { "$" linkTo "${it.first}$${it.second}" }
-				link?.let { typeExpressionStringLinks.add(it) }
-				//val typeExpression = this.value
-				//val link = this.extraValue?.castOrNull<Pair<String, String>>()?.let { "$" linkTo "${it.first}$${it.second}" }
-				//if(typeExpression != null && link != null) {
-				//	info.typeExpressionStringLinks.getOrPut(typeExpression) { SmartList() }.add(link)
-				//}
-			}
 			CwtDataType.FilePath -> {
 				configExpression.value?.let { filePathExpressions.add(it) }
 			}
 			CwtDataType.Icon -> {
 				configExpression.value?.let { iconPathExpressions.add(it) }
+			}
+			CwtDataType.TemplateExpression -> {
+				val link = configExpression.extraValue?.castOrNull<Pair<String, String>>()?.let { "$" linkTo "${it.first}$${it.second}" }
+				link?.let { templateExpressionLinks.add(it) }
+				//val typeExpression = this.value
+				//val link = this.extraValue?.castOrNull<Pair<String, String>>()?.let { "$" linkTo "${it.first}$${it.second}" }
+				//if(typeExpression != null && link != null) {
+				//	info.templateExpressionLinks.getOrPut(typeExpression) { SmartList() }.add(link)
+				//}
 			}
 			else -> pass()
 		}