+++ src/main/kotlin/icu/windea/pls/script/codeInsight/hints/ParadoxDefinitionReferenceInfoHintsProvider.kt
@@ -27,13 +27,13 @@ class ParadoxDefinitionReferenceInfoHintsProvider : ParadoxScriptHintsProvider<N
 		private val settingsKey: SettingsKey<NoSettings> = SettingsKey("ParadoxDefinitionReferenceInfoHintsSettingsKey")
 		private val keyExpressionTypes: Array<CwtDataType> = arrayOf(
 			CwtDataType.TypeExpression,
-			CwtDataType.TypeExpressionString,
+			CwtDataType.TemplateExpression,
 			CwtDataType.AliasName, //需要兼容alias
 			CwtDataType.AliasKeysField //需要兼容alias
 		)
 		private val valueExpressionTypes: Array<CwtDataType> = arrayOf(
 			CwtDataType.TypeExpression,
-			CwtDataType.TypeExpressionString,
+			CwtDataType.TemplateExpression,
 			CwtDataType.SingleAliasRight, //需要兼容single_alias
 			CwtDataType.AliasKeysField, //需要兼容alias
 			CwtDataType.AliasMatchLeft //需要兼容alias